fixbug

+++ routes/api.php
@@ -1,6 +1,7 @@
 <?php
 
 use App\Models\Alilogi\TransportRecharge;
+use App\Models\ExchangeRate;
 use App\Models\OrderItem;
 use App\Models\PurchaseOrder;
 use App\User;
@@ -35,72 +36,90 @@ Route::get('/service_percent', function (Request $request) {
     return $final;
 });
 
+/**
+ * Admin huy don hang cua khach hang
+ */
 Route::post('/cancle-purchase-order', function (Request $request) {
     DB::beginTransaction();
-    $order_id = $request->order_id;
+    // $order_id = $request->order_id;
 
     try {
+        if ($request->ajax()) {
+            $order = PurchaseOrder::find($request->order_id);
+            $status = $order->status;
 
-        $order = PurchaseOrder::find($order_id);
+            // huy don hang moi
+            // chuyen trang thai don hang -> da huy
+            // chuyen trang thai item -> het hang
+            if ($status == PurchaseOrder::STATUS_NEW_ORDER) {
+                $flag = $order->update([
+                    'status'    =>  PurchaseOrder::STATUS_CANCEL,
+                    'purchase_order_service_fee'    =>  0,
+                    'deposit_default'   =>  0,
+                    'purchase_total_items_price'    =>  0,
+                    'final_total_price' =>  0
+                ]);
 
-        $status = $order->status;
+                if ($flag) {
+                    foreach ($order->items as $item) {
+                        $item->qty_reality = 0;
+                        $item->status = OrderItem::STATUS_PURCHASE_OUT_OF_STOCK;
+                        $item->save();
+                    }
 
-        # don hang moi
-        if ($status == PurchaseOrder::STATUS_NEW_ORDER)
-        {
-           # order
-            $order->status = PurchaseOrder::STATUS_CANCEL;
-            $order->save();
-
-            # item
-            foreach ($order->items as $item)
-            {
-                $item->status = OrderItem::STATUS_PURCHASE_OUT_OF_STOCK;
-                $item->save();
+                    DB::commit();
+                    return response()->json([
+                        'error' =>  false,
+                        'msg'   =>  'success'
+                    ]);
+                }
             }
-        } 
-        else if ($status == PurchaseOrder::STATUS_DEPOSITED_ORDERING)
-        {
-            # order
-            $order->status = PurchaseOrder::STATUS_CANCEL;
-            $order->save();
 
-            # item
-            foreach ($order->items as $item)
-            {
-                $item->status = OrderItem::STATUS_PURCHASE_OUT_OF_STOCK;
-                $item->save();
-            }
+            // huy don hang da coc
+            // chuyen trang thai don hang -> da huy
+            // chuyen trang thai item -> het hang
+            // hoan lai tien da coc cho khach
+            // tao giao dich hoan tien
+            else if ($status == PurchaseOrder::STATUS_DEPOSITED_ORDERING) {
+                $deposited = (int) $order->deposited;
 
-            $deposite = $order->deposited;
+                $flag = $order->update([
+                    'status'    =>  PurchaseOrder::STATUS_CANCEL,
+                    'purchase_order_service_fee'    =>  0,
+                    'deposit_default'   =>  0,
+                    'deposited' =>  0,
+                    'deposited_at'  =>  NULL,
+                    'purchase_total_items_price'    =>  0,
+                    'final_total_price' =>  0
+                ]);
 
-            $flag = TransportRecharge::firstOrCreate([
-                'customer_id'   =>  $order->customer_id,
-                'user_id_created'   =>  $request->user_id_created,
-                'money' =>  $deposite,
-                'type_recharge' =>  TransportRecharge::REFUND,
-                'content'   =>  'Hoàn lại tiền cọc đơn hàng ' . $order->order_number
-            ]);
+                if ($flag) {
+                    foreach ($order->items as $item) {
+                        $item->qty_reality = 0;
+                        $item->status = OrderItem::STATUS_PURCHASE_OUT_OF_STOCK;
+                        $item->save();
+                    }
 
-            if ($flag) {
-                $customer = $order->customer;
-                $customer->wallet += $deposite;
-                $customer->save();
+                    $customer = $order->customer;
+                    $customer->wallet += $deposited;
+                    $customer->save();
+
+                    TransportRecharge::firstOrCreate([
+                        'customer_id'   =>  $order->customer_id,
+                        'user_id_created'   =>  $request->user_id_created,
+                        'money' =>  $deposited,
+                        'type_recharge' =>  TransportRecharge::REFUND,
+                        'content'   =>  'Hoàn lại tiền cọc đơn hàng ' . $order->order_number
+                    ]);
+
+                    DB::commit();
+                    return response()->json([
+                        'error' =>  false,
+                        'msg'   =>  'success'
+                    ]);
+                }
             }
         }
-        else {
-            return response()->json([
-                'error' =>  false,
-                'msg'   =>  'Không được phép huỷ đơn hàng này'
-            ]);
-        }
-
-        DB::commit();
-
-        return response()->json([
-            'error' =>  false,
-            'msg'   =>  'success'
-        ]);
     } 
     catch (\Exception $e) {
         DB::rollBack();
@@ -111,6 +130,9 @@ Route::post('/cancle-purchase-order', function (Request $request) {
     }
 });
 
+/**
+ * Admin xac nhan 1 don hang da dat hang tat ca san pham
+ */
 Route::post('/confirm-ordered', function (Request $request) {
     $order_id = $request->order_id;
     $user_id_created = $request->user_id_created;
@@ -118,18 +140,77 @@ Route::post('/confirm-ordered', function (Request $request) {
     DB::beginTransaction();
 
     try {
-        $order = PurchaseOrder::find($order_id);
-        $ordered_items = $order->totalOrderedItems();
-        $total_items = $order->sumQtyRealityItem();
-        if ($ordered_items != $total_items) {
-            return response()->json([
-                'error' =>  true,
-                'msg'   =>  'Đơn hàng này vẫn còn sản phẩm chưa được đặt hàng. Vui lòng xác nhận trạng thái của tất cả sản phẩm trước.'
-            ]);
+        if ($request->ajax()) {
+            $order = PurchaseOrder::find($order_id);
+            $ordered_items = $order->totalOrderedItems();
+            $total_items = $order->sumQtyRealityItem();
+            if ($ordered_items != $total_items) {
+                return response()->json([
+                    'error' =>  true,
+                    'msg'   =>  'Đơn hàng này vẫn còn sản phẩm chưa được đặt hàng. Vui lòng xác nhận trạng thái của tất cả sản phẩm trước.'
+                ]);
+            }
+            else {
+                $order->status = PurchaseOrder::STATUS_ORDERED;
+                $order->user_id_confirm_ordered = $user_id_created;
+                $order->save();
+    
+                DB::commit();
+                return response()->json([
+                    'error' =>  false,
+                    'msg'   =>  'success'
+                ]);
+            }
         }
-        else {
-            $order->status = PurchaseOrder::STATUS_ORDERED;
-            $order->user_id_confirm_ordered = $user_id_created;
+    }
+    catch (\Exception $e) {
+        DB::rollBack();
+        return response()->json([
+            'error' =>  true,
+            'msg'   =>  $e->getMessage()
+        ]);
+    }
+});
+
+/**
+ * Admin xac nhan san pham trong don da het hang
+ */
+Route::post('/confirm-outstock', function (Request $request) {
+    DB::beginTransaction();
+
+    try {
+        if ($request->ajax()) {
+            $item_id = $request->item_id;
+            $item = OrderItem::find($item_id);
+            $order = $item->order;
+
+            $item->qty_reality = 0;
+            $item->status = OrderItem::STATUS_PURCHASE_OUT_OF_STOCK;
+            $item->save();
+
+            $new_data = [
+                'purchase_total_items_price'    =>  0, // tong tien thuc dat
+                'purchase_cn_transport_fee'     =>  0, // tong tien ship noi dia
+                'final_total_price'     =>  0, // tong gia cuoi
+                'deposit_default' => 0, // can coc,
+                'purchase_order_service_fee'    =>  0 // phi dich vu
+            ];
+
+            foreach ($order->items as $item) {
+                if ($item->status != OrderItem::STATUS_PURCHASE_OUT_OF_STOCK) {
+                    $new_data['purchase_cn_transport_fee'] += $item->purchase_cn_transport_fee;
+                    $new_data['purchase_total_items_price'] += $item->qty_reality * $item->price;
+                }
+            }
+
+            $rate = $order->current_rate;
+
+            $percent = (float) PurchaseOrder::PERCENT_NUMBER[$order->customer->customer_percent_service];
+            $new_data['purchase_order_service_fee'] = round($new_data['purchase_total_items_price'] / 100 * $percent, 2);
+            $new_data['final_total_price'] = round( ($new_data['purchase_total_items_price'] ) * $rate);
+            $new_data['deposit_default'] = round($new_data['final_total_price'] * 70 / 100);
+
+            $order->update($new_data);
             $order->save();
 
             DB::commit();
@@ -148,23 +229,74 @@ Route::post('/confirm-ordered', function (Request $request) {
     }
 });
 
-
-Route::post('/confirm-outstock', function (Request $request) {
-    $item_id = $request->item_id;
-
+/**
+ * Khach hang dat coc don hang
+ */
+Route::post('/customer-deposite', function (Request $request) {
     DB::beginTransaction();
-
     try {
-        $item = OrderItem::find($item_id);
-        $item->qty_reality = 0;
-        $item->status = OrderItem::STATUS_PURCHASE_OUT_OF_STOCK;
-        $item->save();
+        if ($request->ajax())
+        {
+            $data = $request->all();
 
-        DB::commit();
-        return response()->json([
-            'error' =>  false,
-            'msg'   =>  'success'
-        ]);
+            $order = PurchaseOrder::find($data['order_id']);
+
+            if ($order->customer->wallet < $order->deposit_default) {
+                return response()->json([
+                    'error' =>  true,
+                    'msg'   =>  'Số dư trong ví của bạn không đủ để thanh toán. Vui lòng liên hệ bộ phận Sale để nạp tiền vào tài khoản.'
+                ]); 
+            }
+            else {
+                $flag_update = PurchaseOrder::find($data['order_id'])
+                ->update([
+                    'deposited' =>  $order->deposit_default,
+                    'user_id_deposited' =>  $order->customer_id,
+                    'deposited_at'  =>  date('Y-m-d', strtotime(now())),
+                    'status'    =>  PurchaseOrder::STATUS_DEPOSITED_ORDERING
+                ]);
+
+                if ($flag_update) {
+                    $flag_deposite = User::find($order->customer_id)
+                    ->update([
+                        'wallet'    =>  $order->customer->wallet - $order->deposit_default
+                    ]);
+
+                    if ($flag_deposite) {
+                        TransportRecharge::firstOrCreate([
+                            'customer_id'   =>  $order->customer_id,
+                            'user_id_created'   => $order->customer_id,
+                            'money' =>  $order->deposit_default,
+                            'type_recharge' =>  TransportRecharge::DEPOSITE_ORDER,
+                            'content'   =>  'Đặt cọc đơn hàng mua hộ. Mã đơn hàng '.$order->order_number,
+                            'order_type'    =>  TransportRecharge::TYPE_ORDER
+                        ]);
+
+
+                        DB::commit();
+                        return response()->json([
+                            'error' =>  false,
+                            'msg'   =>  'success'
+                        ]);
+                    }
+
+                    DB::rollBack();
+                    return response()->json([
+                        'error' =>  false,
+                        'msg'   =>  'success'
+                    ]);
+                }
+
+                DB::rollBack();
+                return response()->json([
+                    'error' =>  false,
+                    'msg'   =>  'success'
+                ]);
+                
+            }
+
+            return $order;
+        }
     }
     catch (\Exception $e) {
         DB::rollBack();
@@ -175,71 +307,22 @@ Route::post('/confirm-outstock', function (Request $request) {
     }
 });
 
-Route::post('/customer-deposite', function (Request $request) {
-    try {
-        $order = PurchaseOrder::find($request->order_id);
-        $customer = $order->customer;
-        $deposited = $order->deposited;
-        $deposite = $order->deposit_default;
-
-        if ($deposited == "") {
-            if ($customer->wallet < $deposited) {
-                return response()->json([
-                    'error' =>  true,
-                    'msg'   =>  'Số dư trong ví của bạn không đủ để thanh toán. Vui lòng liên hệ bộ phận Sale để nạp tiền vào tài khoản.'
-                ]); 
-            }
-            else {
-                $order->deposited = $deposite;
-                $order->user_id_deposited = $order->customer_id;
-                $order->deposited_at =  date('Y-m-d', strtotime(now()));
-                $order->status = PurchaseOrder::STATUS_DEPOSITED_ORDERING;
-                $order->save();
-
-                $customer->wallet -= $deposite;
-                $customer->save();
-
-                TransportRecharge::firstOrCreate([
-                    'customer_id'   =>  $order->customer_id,
-                    'user_id_created'   => $order->customer_id,
-                    'money' =>  $deposite,
-                    'type_recharge' =>  TransportRecharge::DEPOSITE_ORDER,
-                    'content'   =>  'Đặt cọc đơn hàng mua hộ. Mã đơn hàng '.$order->order_number,
-                    'order_type'    =>  TransportRecharge::TYPE_ORDER
-                ]);
-        
-                return response()->json([
-                    'error' =>  false,
-                    'msg'   =>  'success'
-                ]);
-            }
-        }
-
-        return response()->json([
-            'error' =>  true,
-            'msg'   =>  'Đơn hàng này đã đặt cọc.'
-        ]);
-    }
-    catch (\Exception $e) {
-        return response()->json([
-            'error' =>  true,
-            'msg'   =>  $e->getMessage()
-        ]);
-    }
-});
-
+/**
+ * Khách hàng xoá đơn hàng mới
+ */
 Route::post('/customer-destroy', function (Request $request) {
     DB::beginTransaction();
     try {
-        PurchaseOrder::find($request->order_id)->delete();
-        OrderItem::where('order_id', $request->order_id)->delete();
-        DB::commit();
-
-        return response()->json([
-            'error' =>  false,
-            'msg'   =>  'success'
-        ]);
-
+        if ($request->ajax()) {
+            PurchaseOrder::find($request->order_id)->delete();
+            OrderItem::where('order_id', $request->order_id)->delete();
+            DB::commit();
+    
+            return response()->json([
+                'error' =>  false,
+                'msg'   =>  'success'
+            ]);
+        }
     }
     catch (\Exception $e) {
         DB::rollBack();
@@ -272,50 +355,63 @@ Route::post('/customer-delete-item-from-cart', function (Request $request) {
     }
 });
 
-
+/**
+ * Khách hàng xoá sản phẩm khỏi đơn hàng
+ */
 Route::post('/customer-delete-item-from-order', function (Request $request) {
-    
+    DB::beginTransaction();
     try {
-        $item = OrderItem::find($request->id);
-        if ($item) {
-            $order = PurchaseOrder::find($item->order_id);
+        if ($request->ajax()) {
+            $item = OrderItem::find($request->id);
+            $order_id = $item->order_id;
+            $order = PurchaseOrder::find($order_id);
 
-            $item->delete();
-    
-            $current_items = $order->items;
-            $res = [
-                'purchase_total_items_price'    =>  0,
-                'purchase_cn_transport_fee'     =>  0,
-                'final_total_price'     =>  0,
-                'deposit_default' => 0
-            ];
-    
-            if ($current_items->count() > 0)
-            {
-                $purchase_total_items_price = 0; // tong tien sp
-                $purchase_cn_transport_fee = 0; // tong tien van chuyen
-                foreach ($current_items as $current_item) {
-                    if ($current_item->status != OrderItem::STATUS_PURCHASE_OUT_OF_STOCK) {
-                        $res['purchase_cn_transport_fee'] += $item->purchase_cn_transport_fee;
-                        $res['purchase_total_items_price'] += $item->qty_reality * $item->price;
+            if ($order->items->count() == 1) {
+                return response()->json([
+                    'error' =>  true,
+                    'msg'   =>  'Đơn hàng này có duy nhất 1 sản phẩm. Bạn không thể xoá sản phẩm này khỏi đơn hàng.'
+                ]);
+            }
+
+            $flag = $item->delete();
+
+            if ($flag) {
+                $new_data = [
+                    'purchase_total_items_price'    =>  0, // tong tien thuc dat
+                    'purchase_cn_transport_fee'     =>  0, // tong tien ship noi dia
+                    'final_total_price'     =>  0, // tong gia cuoi
+                    'deposit_default' => 0, // can coc,
+                    'purchase_order_service_fee'    =>  0 // phi dich vu
+                ];
+
+                foreach ($order->items as $item) {
+                    if ($item->status != OrderItem::STATUS_PURCHASE_OUT_OF_STOCK) {
+                        $new_data['purchase_cn_transport_fee'] += $item->purchase_cn_transport_fee;
+                        $new_data['purchase_total_items_price'] += $item->qty_reality * $item->price;
                     }
                 }
-        
-                $res['final_total_price'] = ($res['purchase_total_items_price'] + $res['purchase_cn_transport_fee'] + $order->purchase_order_service_fee);
-                $res['deposit_default'] = $res['final_total_price'] * 70 / 100;
+
+                $rate = ExchangeRate::first()->vnd;
+
+                $percent = (float) PurchaseOrder::PERCENT_NUMBER[$order->customer->customer_percent_service];
+                $new_data['purchase_order_service_fee'] = round($new_data['purchase_total_items_price'] / 100 * $percent, 2);
+                $new_data['final_total_price'] = round( ($new_data['purchase_total_items_price'] ) * $rate);
+                $new_data['deposit_default'] = round($new_data['final_total_price'] * 70 / 100);
+
+                $order->update($new_data);
+                $order->save();
+                
+                DB::commit();
+                return response()->json([
+                    'error' =>  false,
+                    'msg'   =>  'success'
+                ]);
             }
-            
-            $order->update($res);
-            $order->save();
-    
-            return response()->json([
-                'error' =>  false,
-                'msg'   =>  'success'
-            ]);
         }
+        
     }
     catch (\Exception $e) {
-        
+        DB::rollBack();
         return response()->json([
             'error' =>  true,
             'msg'   =>  $e->getMessage()
